Refactor: Move blocking map to pixels-core/utils
